@@ -1,4 +1,4 @@
-class Permission < ActiveRecord::Base  
+class Permission < ApplicationRecord
   belongs_to :user
   belongs_to :role
 end
