@@ -1,4 +1,4 @@
-class Role < ActiveRecord::Base
+class Role < ApplicationRecord
 
   has_many :permissions, :dependent => :destroy
   has_many :users, :through => :permissions
