@@ -1,10 +1,10 @@
-class Api::SessionsController < ApplicationController
-  
+class Api::JwtsController < ApplicationController
+
   # Require our abstraction for encoding/deconding JWT.
   require 'auth_token'
 
   before_action :validate_token
-  
+
   respond_to :json
 
   def show
